@@ -1,52 +1,67 @@
 name: 'example'
 description: 'Example application for the JSON Theme'
 publish_to: 'none'
-version: '1.0.0+36'
+version: '1.0.0+39'
 
-environment: 
-  sdk: '>=3.0.0 <4.0.0'
+environment:
+  sdk: '>=3.3.0 <4.0.0'
 
-dependencies: 
-  flutter: 
+dependencies:
+  flutter:
     sdk: 'flutter'
-  form_validation: '^3.0.2+2'
-  google_fonts: '^5.1.0'
-  intl: '^0.18.1'
-  json_theme: 
+  form_validation: '^3.1.1+5'
+  google_fonts: '^6.2.1'
+  intl: '^0.19.0'
+  json_theme:
     path: '../'
-  meta: '^1.9.1'
+  meta: '^1.12.0'
 
-dev_dependencies: 
-  flutter_test: 
+dev_dependencies:
+  flutter_lints: '^4.0.0'
+  flutter_test:
     sdk: 'flutter'
 
-
-flutter: 
+flutter:
   uses-material-design: true
-  assets: 
+  assets:
     - 'assets/themes/'
-  fonts: 
-    - 
-      family: 'lato'
-      fonts: 
-        - 
-          asset: 'assets/fonts/Lato-Regular.ttf'
+  fonts:
+    - family: 'lato'
+      fonts:
+        - asset: 'assets/fonts/Lato-Regular.ttf'
+    - family: 'metal'
+      fonts:
+        - asset: 'assets/fonts/MetalMania-Regular.ttf'
+    - family: 'MaterialIcons'
+      fonts:
+        - asset: 'assets/fonts/MaterialIcons-Regular.ttf'
 
-    - 
-      family: 'metal'
-      fonts: 
-        - 
-          asset: 'assets/fonts/MetalMania-Regular.ttf'
+permittedLicenses:
+  - Apache-2.0
+  - BSD-2-Clause
+  - BSD-3-Clause
+  - MIT
+  - MIT-Modern-Variant
+  - MPL-2.0
+  - Zlib
 
-    - 
-      family: 'MaterialIcons'
-      fonts: 
-        - 
-          asset: 'assets/fonts/MaterialIcons-Regular.ttf'
+# The [license_checker](https://pub.dev/packages/license_checker) package cannot
+# detect the built in Flutter packages because they aren't published to pub.dev
+# and do not have a LICENSE file in their respective folders.  So this section
+# informs the license_checker that all the built in Flutter packages all have
+# the same BSD-3-Clause license as Flutter itself. 
+packageLicenseOverride: 
+  flutter: BSD-3-Clause
+  flutter_driver: BSD-3-Clause
+  flutter_goldens: BSD-3-Clause
+  flutter_localizations: BSD-3-Clause
+  flutter_web_plugins: BSD-3-Clause
+  flutter_test: BSD-3-Clause
+  fuchsia_remote_debug_protocol: BSD-3-Clause
+  integration_test: BSD-3-Clause
+  rxdart: Apache-2.0
 
-
-
-ignore_updates: 
+ignore_updates:
   - 'archive'
   - 'async'
   - 'boolean_selector'
@@ -69,5 +84,6 @@ ignore_updates:
   - 'term_glyph'
   - 'test_api'
   - 'typed_data'
+  - 'uuid'
   - 'vector_math'
   - 'webdriver'
