@@ -1,65 +1,85 @@
-name: 'json_theme'
-description: 'A library to dynamically generate a ThemeData object from a JSON file or dynamic map object'
-homepage: 'https://github.com/peiffer-innovations/json_theme'
-version: '6.3.1'
+name: "json_theme"
+description: "A library to dynamically generate a ThemeData object from a JSON file or dynamic map object"
+homepage: "https://github.com/peiffer-innovations/json_theme"
+version: "6.5.0+1"
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 
 analyzer:
   exclude:
-    - 'lib/generated/**'
-    - 'lib/**/*.g.dart'
+    - "lib/generated/**"
+    - "lib/**/*.g.dart"
 
 dependencies:
   flutter:
-    sdk: 'flutter'
-  json_class: '^3.0.0'
-  json_schema: '^5.1.3'
-  json_theme_annotation: '^1.0.1+1'
-  logging: '^1.2.0'
-  meta: '^1.9.1'
+    sdk: "flutter"
+  json_class: "^3.0.0+13"
+  json_schema: "^5.1.7"
+  json_theme_annotation: "^1.0.3+6"
+  logging: "^1.2.0"
+  meta: "^1.12.0"
 
 dev_dependencies:
-  analyzer: '^6.2.0'
-  build: '^2.4.1'
-  build_runner: '^2.4.6'
-  code_builder: '^4.7.0'
-  flutter_lints: '^2.0.2'
+  analyzer: "^6.4.1"
+  build: "^2.4.1"
+  build_runner: "^2.4.9"
+  code_builder: "^4.10.0"
+  flutter_lints: "^4.0.0"
   flutter_test:
-    sdk: 'flutter'
-  json_theme_codegen: '^1.1.0'
-  recase: '^4.1.0'
-  source_gen: '^1.4.0'
+    sdk: "flutter"
+  json_theme_codegen: "^1.1.2+13"
+  recase: "^4.1.0"
+  source_gen: "^1.5.0"
 
-# dependency_overrides:
-#   json_theme_annotation:
-#     path: ../annotation
-#   json_theme_codegen:
-#     path: ../codegen
+permittedLicenses:
+  - Apache-2.0
+  - BSD-2-Clause
+  - BSD-3-Clause
+  - MIT
+  - MIT-Modern-Variant
+  - MPL-2.0
+  - Zlib
+
+# The [license_checker](https://pub.dev/packages/license_checker) package cannot
+# detect the built in Flutter packages because they aren't published to pub.dev
+# and do not have a LICENSE file in their respective folders.  So this section
+# informs the license_checker that all the built in Flutter packages all have
+# the same BSD-3-Clause license as Flutter itself.
+packageLicenseOverride:
+  flutter: BSD-3-Clause
+  flutter_driver: BSD-3-Clause
+  flutter_goldens: BSD-3-Clause
+  flutter_localizations: BSD-3-Clause
+  flutter_web_plugins: BSD-3-Clause
+  flutter_test: BSD-3-Clause
+  fuchsia_remote_debug_protocol: BSD-3-Clause
+  integration_test: BSD-3-Clause
+  rxdart: Apache-2.0
 
 ignore_updates:
-  - 'archive'
-  - 'async'
-  - 'boolean_selector'
-  - 'characters'
-  - 'charcode'
-  - 'collection'
-  - 'clock'
-  - 'crypto'
-  - 'fake_async'
-  - 'file'
-  - 'matcher'
-  - 'material_color_utilities'
-  - 'meta'
-  - 'path'
-  - 'source_span'
-  - 'stack_trace'
-  - 'stream_channel'
-  - 'string_scanner'
-  - 'sync_http'
-  - 'term_glyph'
-  - 'test_api'
-  - 'typed_data'
-  - 'vector_math'
-  - 'webdriver'
+  - "archive"
+  - "async"
+  - "boolean_selector"
+  - "characters"
+  - "charcode"
+  - "collection"
+  - "clock"
+  - "crypto"
+  - "fake_async"
+  - "file"
+  - "matcher"
+  - "material_color_utilities"
+  - "meta"
+  - "path"
+  - "source_span"
+  - "stack_trace"
+  - "stream_channel"
+  - "string_scanner"
+  - "sync_http"
+  - "term_glyph"
+  - "test_api"
+  - "typed_data"
+  - "uuid"
+  - "vector_math"
+  - "webdriver"
