@@ -16,11 +16,13 @@ class PopupMenuThemeDataSchema {
       'color': SchemaHelper.objectSchema(ColorSchema.id),
       'elevation': SchemaHelper.numberSchema,
       'enableFeedback': SchemaHelper.boolSchema,
+      'iconColor': SchemaHelper.objectSchema(ColorSchema.id),
+      'iconSize': SchemaHelper.numberSchema,
       'labelTextStyle': SchemaHelper.objectSchema(
-        MaterialStatePropertyTextStyleSchema.id,
+        WidgetStatePropertyTextStyleSchema.id,
       ),
       'mouseCursor': SchemaHelper.objectSchema(
-        MaterialStatePropertyMouseCursorSchema.id,
+        WidgetStatePropertyMouseCursorSchema.id,
       ),
       'position': SchemaHelper.objectSchema(PopupMenuPositionSchema.id),
       'shadowColor': SchemaHelper.objectSchema(ColorSchema.id),
