@@ -15,6 +15,9 @@ class SliderThemeDataSchema {
     'properties': {
       'activeTickMarkColor': SchemaHelper.objectSchema(ColorSchema.id),
       'activeTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
+      'allowedInteraction': SchemaHelper.objectSchema(
+        SliderInteractionSchema.id,
+      ),
       'disabledActiveTickMarkColor': SchemaHelper.objectSchema(ColorSchema.id),
       'disabledActiveTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
       'disabledInactiveTickMarkColor': SchemaHelper.objectSchema(
@@ -29,7 +32,7 @@ class SliderThemeDataSchema {
       'inactiveTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
       'minThumbSeparation': SchemaHelper.numberSchema,
       'mouseCursor': SchemaHelper.objectSchema(
-        MaterialStatePropertyMouseCursorSchema.id,
+        WidgetStatePropertyMouseCursorSchema.id,
       ),
       'overlappingShapeStrokeColor': SchemaHelper.objectSchema(ColorSchema.id),
       'overlayColor': SchemaHelper.objectSchema(ColorSchema.id),
