@@ -1,15 +1,12 @@
 import 'package:json_theme/json_theme_schemas.dart';
 
-class MaterialStatePropertyBoolSchema {
+class WidgetStatePropertyBoolSchema {
   static const id =
-      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_theme/material_state_property_bool.json';
+      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_theme/widget_state_property_bool.json';
 
   static final schema = {
     r'$schema': 'http://json-schema.org/draft-07/schema#',
     r'$id': id,
-    r'$comment':
-        'https://api.flutter.dev/flutter/material/MaterialStateProperty-class.html',
-    'title': 'MaterialStatePropertyBool',
     'type': 'object',
     'properties': {
       'disabled': SchemaHelper.boolSchema,
