@@ -1,14 +1,14 @@
 import 'package:json_theme/json_theme_schemas.dart';
 
-class MaterialStatePropertyIconSchema {
+class WidgetStatePropertyIconSchema {
   static const id =
-      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_theme/material_state_property_icon.json';
+      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_theme/widget_state_property_icon.json';
 
   static final schema = {
     r'$schema': 'http://json-schema.org/draft-07/schema#',
     r'$id': id,
     r'$comment': 'https://api.flutter.dev/flutter/widgets/Icon-class.html',
-    'title': 'MaterialStatePropertyIcon',
+    'title': 'WidgetStatePropertyIcon',
     'type': 'object',
     'properties': {
       'disabled': SchemaHelper.objectSchema(IconSchema.id),
