@@ -1,15 +1,13 @@
 import 'package:json_theme/json_theme_schemas.dart';
 
-class MaterialStatePropertyTextStyleSchema {
+class WidgetStatePropertyTextStyleSchema {
   static const id =
-      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_theme/material_state_property_text_style.json';
+      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_theme/widget_state_property_text_style.json';
 
   static final schema = {
     r'$schema': 'http://json-schema.org/draft-07/schema#',
     r'$id': id,
-    r'$comment':
-        'https://api.flutter.dev/flutter/material/MaterialStateProperty-class.html',
-    'title': 'MaterialStatePropertyTextStyle',
+    'title': 'WidgetStatePropertyTextStyle',
     'type': 'object',
     'properties': {
       'disabled': SchemaHelper.objectSchema(TextStyleSchema.id),
