@@ -2,4 +2,4 @@ export 'json_theme_schemas.dart';
 //
 export 'src/codec/theme_decoder.dart';
 export 'src/codec/theme_encoder.dart';
-export 'src/model/json_material_state_property.dart';
+export 'src/model/json_widget_state_property.dart';
