@@ -40,6 +40,7 @@ export 'schemas/enums/max_length_enforcement_schema.dart';
 export 'schemas/enums/navigation_destination_label_behavior_schema.dart';
 export 'schemas/enums/navigation_rail_label_type_schema.dart';
 export 'schemas/enums/notched_shape_schema.dart';
+export 'schemas/enums/overflow_box_fit_schema.dart';
 export 'schemas/enums/page_transitions_builder_schema.dart';
 export 'schemas/enums/pan_axis_schema.dart';
 export 'schemas/enums/pointer_device_kind_schema.dart';
@@ -50,11 +51,13 @@ export 'schemas/enums/scroll_view_keyboard_dismiss_behavior_schema.dart';
 export 'schemas/enums/scrollbar_orientation_schema.dart';
 export 'schemas/enums/show_value_indicator_schema.dart';
 export 'schemas/enums/slider_component_shape_schema.dart';
+export 'schemas/enums/slider_interaction_schema.dart';
 export 'schemas/enums/slider_tick_mark_shape_schema.dart';
 export 'schemas/enums/smart_dashes_type_schema.dart';
 export 'schemas/enums/smart_quotes_type_schema.dart';
 export 'schemas/enums/snack_bar_behavior_schema.dart';
 export 'schemas/enums/stack_fit_schema.dart';
+export 'schemas/enums/tab_alignment_schema.dart';
 export 'schemas/enums/tab_bar_indicator_size_schema.dart';
 export 'schemas/enums/target_platform_schema.dart';
 export 'schemas/enums/text_align_schema.dart';
@@ -76,6 +79,7 @@ export 'schemas/enums/visual_density_schema.dart';
 export 'schemas/enums/wrap_alignment_schema.dart';
 export 'schemas/enums/wrap_cross_alignment_schema.dart';
 //
+export 'schemas/objects/action_icon_theme_data_schema.dart';
 export 'schemas/objects/alignment_geometry_schema.dart';
 export 'schemas/objects/alignment_schema.dart';
 export 'schemas/objects/app_bar_theme_schema.dart';
@@ -130,17 +134,6 @@ export 'schemas/objects/list_tile_theme_data_schema.dart';
 export 'schemas/objects/locale_schema.dart';
 export 'schemas/objects/material_banner_theme_data_schema.dart';
 export 'schemas/objects/material_color_schema.dart';
-export 'schemas/objects/material_state_property_bool_schema.dart';
-export 'schemas/objects/material_state_property_border_side_schema.dart';
-export 'schemas/objects/material_state_property_color_schema.dart';
-export 'schemas/objects/material_state_property_double_schema.dart';
-export 'schemas/objects/material_state_property_edge_insets_geometry_schema.dart';
-export 'schemas/objects/material_state_property_icon_schema.dart';
-export 'schemas/objects/material_state_property_icon_theme_data_schema.dart';
-export 'schemas/objects/material_state_property_mouse_cursor_schema.dart';
-export 'schemas/objects/material_state_property_outlined_border_schema.dart';
-export 'schemas/objects/material_state_property_size_schema.dart';
-export 'schemas/objects/material_state_property_text_style_schema.dart';
 export 'schemas/objects/matrix4_schema.dart';
 export 'schemas/objects/menu_bar_theme_data_schema.dart';
 export 'schemas/objects/menu_button_theme_data_schema.dart';
@@ -192,3 +185,15 @@ export 'schemas/objects/time_picker_theme_data_schema.dart';
 export 'schemas/objects/toggle_buttons_theme_data_schema.dart';
 export 'schemas/objects/tooltip_theme_data_schema.dart';
 export 'schemas/objects/typography_schema.dart';
+export 'schemas/objects/widget_state_color_schema.dart';
+export 'schemas/objects/widget_state_property_bool_schema.dart';
+export 'schemas/objects/widget_state_property_border_side_schema.dart';
+export 'schemas/objects/widget_state_property_color_schema.dart';
+export 'schemas/objects/widget_state_property_double_schema.dart';
+export 'schemas/objects/widget_state_property_edge_insets_geometry_schema.dart';
+export 'schemas/objects/widget_state_property_icon_schema.dart';
+export 'schemas/objects/widget_state_property_icon_theme_data_schema.dart';
+export 'schemas/objects/widget_state_property_mouse_cursor_schema.dart';
+export 'schemas/objects/widget_state_property_outlined_border_schema.dart';
+export 'schemas/objects/widget_state_property_size_schema.dart';
+export 'schemas/objects/widget_state_property_text_style_schema.dart';
