@@ -34,12 +34,13 @@ class ColorSchemeSchema {
       'outlineVariant': SchemaHelper.objectSchema(ColorSchema.id),
       'primary': SchemaHelper.objectSchema(ColorSchema.id),
       'primaryContainer': SchemaHelper.objectSchema(ColorSchema.id),
+      'scrim': SchemaHelper.objectSchema(ColorSchema.id),
       'secondary': SchemaHelper.objectSchema(ColorSchema.id),
       'secondaryContainer': SchemaHelper.objectSchema(ColorSchema.id),
       'shadow': SchemaHelper.objectSchema(ColorSchema.id),
       'surface': SchemaHelper.objectSchema(ColorSchema.id),
       'surfaceTint': SchemaHelper.objectSchema(ColorSchema.id),
-      'surfaceVariant': SchemaHelper.objectSchema(ColorSchema.id),
+      'JsonWidgetStateProperty': SchemaHelper.numberSchema,
       'tertiary': SchemaHelper.objectSchema(ColorSchema.id),
       'tertiaryContainer': SchemaHelper.objectSchema(ColorSchema.id)
     },
